integrate frontend with backend

+++ LabResultsApi/DTOs/FileDataTestDto.cs
@@ -4,7 +4,7 @@ public class FileDataTestDto
 {
     public int SampleId { get; set; }
     public short TestId { get; set; }
-    public int TrialNumber { get; set; }
+    public short TrialNumber { get; set; }
     public decimal Value1 { get; set; }
     public string FileContent { get; set; } = string.Empty;
 }
